fix: #108 📌 rename dir
