+++ book/impls/50_basic_template_compiler/010_transformer/packages/compiler-core/ast.ts
@@ -1,5 +1,5 @@
 import { isString } from "../shared";
-import { PropsExpression } from "./transform/transformElement";
+import { PropsExpression } from "./transforms/transformElement";
 
 export const enum NodeTypes {
   ROOT,
