+++ book/impls/50_basic_template_compiler/010_transformer/packages/compiler-core/compile.ts
@@ -2,7 +2,7 @@ import { generate } from "./codegen";
 import { CompilerOptions } from "./options";
 import { baseParse } from "./parse";
 import { DirectiveTransform, NodeTransform, transform } from "./transform";
-import { transformElement } from "./transform/transformElement";
+import { transformElement } from "./transforms/transformElement";
 
 export type TransformPreset = [
   NodeTransform[],
