+++ book/impls/50_basic_template_compiler/020_v_bind/packages/compiler-core/transforms/vBind.ts
