+++ book/impls/50_basic_template_compiler/025_v_on/packages/compiler-core/ast.ts
@@ -1,7 +1,7 @@
 import { isString } from "../shared";
 import { CREATE_VNODE } from "./runtimeHelpers";
 import { TransformContext } from "./transform";
-import { PropsExpression } from "./transform/transformElement";
+import { PropsExpression } from "./transforms/transformElement";
 
 export const enum NodeTypes {
   ROOT,
