+++ book/impls/50_basic_template_compiler/025_v_on/packages/compiler-core/compile.ts
@@ -2,9 +2,9 @@ import { generate } from "./codegen";
 import { CompilerOptions } from "./options";
 import { baseParse } from "./parse";
 import { DirectiveTransform, NodeTransform, transform } from "./transform";
-import { transformElement } from "./transform/transformElement";
-import { transformBind } from "./transform/vBind";
-import { transformOn } from "./transform/vOn";
+import { transformElement } from "./transforms/transformElement";
+import { transformBind } from "./transforms/vBind";
+import { transformOn } from "./transforms/vOn";
 
 export type TransformPreset = [
   NodeTransform[],
