+++ book/impls/50_basic_template_compiler/025_v_on/packages/compiler-core/transform.ts
@@ -19,8 +19,7 @@ export type NodeTransform = (
 export type DirectiveTransform = (
   dir: DirectiveNode,
   node: ElementNode,
-  context: TransformContext,
-  augmentor?: (ret: DirectiveTransformResult) => DirectiveTransformResult
+  context: TransformContext
 ) => DirectiveTransformResult;
 
 export interface DirectiveTransformResult {
